requirements.txt
asgiref==3.8.1
Brotli==1.1.0
click==8.1.7
colorama==0.4.6
dj-database-url==2.2.0
Django==5.1.2
djangorestframework==3.15.2
gunicorn==23.0.0
h11==0.14.0
packaging==24.1
psycopg2-binary==2.9.9
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.32.0
whitenoise==6.7.0
